+++ lib/Exif.php
@@ -68,9 +68,15 @@ class Exif
     /**
      * Get exif data as a JSON object from a Nextcloud file.
      */
-    public static function getExifFromFile(File &$file)
+    public static function getExifFromFile(File $file)
     {
-        $path = $file->getStorage()->getLocalFile($file->getInternalPath());
+        try {
+            $path = $file->getStorage()->getLocalFile($file->getInternalPath());
+        } catch (\Throwable $ex) {
+            // https://github.com/pulsejet/memories/issues/820
+            throw new \Exception('Failed to get local file: '.$ex->getMessage());
+        }
+
         if (!\is_string($path)) {
             throw new \Exception('Failed to get local file path');
         }
@@ -101,7 +107,7 @@ class Exif
     }
 
     /** Get exif data as a JSON object from a local file path */
-    public static function getExifFromLocalPath(string &$path)
+    public static function getExifFromLocalPath(string $path)
     {
         if (null !== self::$staticProc) {
             self::ensureStaticExiftoolProc();
@@ -229,7 +235,7 @@ class Exif
      *
      * @return array [width, height]
      */
-    public static function getDimensions(array &$exif)
+    public static function getDimensions(array $exif)
     {
         $width = $exif['ImageWidth'] ?? 0;
         $height = $exif['ImageHeight'] ?? 0;
@@ -397,7 +403,7 @@ class Exif
         return $buf;
     }
 
-    private static function getExifFromLocalPathWithStaticProc(string &$path)
+    private static function getExifFromLocalPathWithStaticProc(string $path)
     {
         $args = implode("\n", self::EXIFTOOL_ARGS);
         fwrite(self::$staticPipes[0], "{$path}\n{$args}\n-execute\n");
@@ -419,7 +425,7 @@ class Exif
         }
     }
 
-    private static function getExifFromLocalPathWithSeparateProc(string &$path, array $extraArgs = [])
+    private static function getExifFromLocalPathWithSeparateProc(string $path, array $extraArgs = [])
     {
         $pipes = [];
         $proc = proc_open(array_merge(self::getExiftool(), self::EXIFTOOL_ARGS, $extraArgs, [$path]), [
@@ -445,7 +451,7 @@ class Exif
     }
 
     /** Get json array from stdout of exiftool */
-    private static function processStdout(string &$stdout)
+    private static function processStdout(string $stdout)
     {
         $json = json_decode($stdout, true);
         if (!$json) {