+++ src/App.vue
@@ -54,10 +54,13 @@
 
     <Viewer />
     <Sidebar />
+
     <EditMetadataModal />
     <AddToAlbumModal />
     <NodeShareModal />
     <ShareModal />
+    <MoveToFolderModal />
+    <FaceMoveModal />
   </NcContent>
 </template>
 
@@ -83,12 +86,15 @@ import FirstStart from './components/FirstStart.vue';
 import Viewer from './components/viewer/Viewer.vue';
 import Metadata from './components/Metadata.vue';
 import Sidebar from './components/Sidebar.vue';
+import MobileNav from './components/MobileNav.vue';
+import MobileHeader from './components/MobileHeader.vue';
+
 import EditMetadataModal from './components/modal/EditMetadataModal.vue';
 import AddToAlbumModal from './components/modal/AddToAlbumModal.vue';
 import NodeShareModal from './components/modal/NodeShareModal.vue';
 import ShareModal from './components/modal/ShareModal.vue';
-import MobileNav from './components/MobileNav.vue';
-import MobileHeader from './components/MobileHeader.vue';
+import MoveToFolderModal from './components/modal/MoveToFolderModal.vue';
+import FaceMoveModal from './components/modal/FaceMoveModal.vue';
 
 import ImageMultiple from 'vue-material-design-icons/ImageMultiple.vue';
 import FolderIcon from 'vue-material-design-icons/Folder.vue';
@@ -123,12 +129,15 @@ export default defineComponent({
     FirstStart,
     Viewer,
     Sidebar,
+    MobileNav,
+    MobileHeader,
+
     EditMetadataModal,
     AddToAlbumModal,
     NodeShareModal,
     ShareModal,
-    MobileNav,
-    MobileHeader,
+    MoveToFolderModal,
+    FaceMoveModal,
 
     ImageMultiple,
     FolderIcon,
@@ -240,7 +249,7 @@ export default defineComponent({
     utils.bus.on('memories:user-config-changed', this.refreshNav);
 
     // Register global functions
-    globalThis.showSettings = () => this.showSettings();
+    mModals.showSettings = this.showSettings;
   },
 
   mounted() {