+++ src/components/modal/FaceMoveModal.vue
@@ -41,17 +41,15 @@ export default defineComponent({
     FaceList,
   },
 
-  props: {
-    updateLoading: {
-      type: Function,
-      required: true,
-    },
-  },
-
   data: () => ({
     show: false,
     photos: [] as IPhoto[],
   }),
+
+  created() {
+    console.assert(!mModals.moveToFace, 'FaceMoveModal created twice');
+    mModals.moveToFace = this.open;
+  },
 
   methods: {
     open(photos: IPhoto[]) {
@@ -81,8 +79,8 @@ export default defineComponent({
       this.$emit('close');
     },
 
-    moved(list: IPhoto[]) {
-      this.$emit('moved', list);
+    moved(photos: IPhoto[]) {
+      utils.bus.emit('memories:timeline:deleted', photos);
     },
 
     async clickFace(face: IFace) {
@@ -106,7 +104,6 @@ export default defineComponent({
 
       try {
         this.show = false;
-        this.updateLoading(1);
 
         // Create map to return IPhoto later
         const map = new Map<number, IPhoto>();
@@ -123,7 +120,6 @@ export default defineComponent({
         console.error(error);
         showError(this.t('photos', 'An error occurred while moving photos from {name}.', { name }));
       } finally {
-        this.updateLoading(-1);
         this.close();
       }
     },