+++ src/components/modal/MoveToFolderModal.vue
@@ -41,6 +41,11 @@ export default defineComponent({
     processing: false,
   }),
 
+  created() {
+    console.assert(!mModals.moveToFolder, 'MoveToFolderModal created twice');
+    mModals.moveToFolder = this.open;
+  },
+
   methods: {
     open(photos: IPhoto[]) {
       this.photosDone = 0;
@@ -50,14 +55,9 @@ export default defineComponent({
       this.chooseFolderPath();
     },
 
-    moved(photos: IPhoto[]) {
-      this.$emit('moved', photos);
-    },
-
     close() {
       this.photos = [];
       this.processing = false;
-      this.$emit('close');
     },
 
     async chooseFolderPath() {
@@ -72,7 +72,7 @@ export default defineComponent({
 
       for await (const fids of gen) {
         this.photosDone += fids.filter((f) => f).length;
-        this.moved(this.photos.filter((p) => fids.includes(p.fileid)));
+        utils.bus.emit('memories:timeline:soft-refresh', null);
       }
 
       const n = this.photosDone;